Moved BasicTabs function into its own file.
